- const size_t& -> UINT로 수정하여 타입 캐스팅 최소화 - 함수 이름 수정

+++ D3D11Manager/D3D11Manager/AShader.cpp
@@ -28,5 +28,5 @@ void D3D11::AShader::SetSamplerState(
 	const vector<ID3D11SamplerState*>& samplerStates
 ) noexcept
 {
-	m_samplerStateCount = samplerStates.size();
+	m_samplerStateCount = static_cast<UINT>(samplerStates.size());
 }
